Detection if point is inside of the object added.
Detection if point is inside of the object added. This is used for monte
carlo method to compute volume of the object.

For given point we create vector in a random direction. Then we check
each triangle for intersection with the vector. Based on the number of
triangles intersected, the point is assigned position relative to the
object. (1 intersection means inside of the object, 2 means outside, 3
for inside etc..)